+++ Unit Test/GraphicsTest.h
@@ -1,0 +1,10 @@
+#pragma once
+
+
+#include "EventHandlerTest.h"
+
+void GrahpicsTest() {
+	EventHanlderTest();
+
+
+}
